rename `WidgetState` to `MapState`
