+++ lib/src/platform/platform_native.dart
@@ -3,9 +3,9 @@ import 'dart:io';
 import 'package:maplibre/src/map_state.dart';
 import 'package:maplibre/src/offline/offline_manager.dart';
 import 'package:maplibre/src/permission_manager.dart';
+import 'package:maplibre/src/platform/android/map_state.dart';
 import 'package:maplibre/src/platform/android/offline_manager.dart';
 import 'package:maplibre/src/platform/android/permission_manager.dart';
-import 'package:maplibre/src/platform/android/widget_state.dart';
 import 'package:maplibre/src/platform_interface.dart';
 
 /// An implementation of [PlatformInterface] that uses method channels and ffi.
