+++ lib/src/platform/web/map_state.dart
